fastapi-app/requirements.txt
aiosqlite==0.16.0
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.25.0
bcrypt==3.2.0
cffi==1.14.4
click==8.1.3
deta==1.1.0
fastapi==0.63.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
httptools==0.4.0
Hypercorn==0.11.1
hyperframe==6.0.0
idna==3.3
iso8601==0.1.13
passlib==1.7.4
priority==1.3.0
pycparser==2.20
pydantic==1.7.3
PyJWT==2.0.0
PyPika==0.44.1
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2020.5
PyYAML==6.0
six==1.15.0
sniffio==1.2.0
starlette==0.13.6
toml==0.10.2
tortoise-orm==0.16.19
typing-extensions==3.7.4.3
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
wsproto==1.0.0
